Crate rtriangulate

A Rust implementation of the Delaunay triangulation algorithm presented by Paul Bourke.

Usage

Add the rtriangulate dependency to Cargo.toml:

[dependencies]
rtriangulate = "0.2"

And use the crate as such:

extern crate rtriangulate;

use rtriangulate::{Point, Triangle, triangulate};

// A list of points (which has to be sorted on x).
let points = [Point::new(10.0, 50.0), Point::new(25.0, 40.0), Point::new(30.0, 40.0)];
let triangles = triangulate(&points).unwrap();

assert_eq!(triangles, [Triangle(1, 0, 2)]);

Structs

Edge
An edge, represented by indexes into a list of points.

Point
A two-dimensional point.

Triangle
A triangle, represented by indexes into a list of points.

Enums

TriangulateError
Possible triangulation errors.

Functions

triangulate
Generate the Delaunay triangulation of given set of points.

Type Definitions

Result